+++ crates/gen/src/types/mod.rs
@@ -2,7 +2,6 @@ pub use super::*;
 
 mod bool32;
 mod bstr;
-mod callback;
 mod class;
 mod com_interface;
 mod delegate;
@@ -22,7 +21,6 @@ mod vector4;
 
 pub use bool32::*;
 pub use bstr::*;
-pub use callback::*;
 pub use class::*;
 pub use com_interface::*;
 pub use delegate::*;
